% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{combinedTable,metabCombiner-method}
\alias{combinedTable,metabCombiner-method}
\title{Obtain metabCombiner Report}
\usage{
\S4method{combinedTable}{metabCombiner}(object)
}
\arguments{
\item{object}{metabCombiner object.}
}
\value{
data.frame \code{combinedTable} merged feature alignment report. The
columns of the report are as follows:

\item{idx}{Identities of features from dataset X}
\item{idy}{Identities of features from dataset Y}
\item{mzx}{m/z values of features from dataset X}
\item{mzy}{m/z values of features from dataset Y}
\item{rtx}{retention time values of features from dataset X}
\item{rty}{retention time values of features from dataset Y}
\item{rtProj}{model-projected retention time values from X to Y}
\item{Qx}{abundance quantile values of features from dataset X}
\item{Qy}{abundance quantile values of features from dataset Y}
\item{group}{m/z feature group}
\item{score}{computed similarity scores of feature pairing}
\item{rankX}{ranking of pairing score for X dataset features}
\item{rankY}{ranking of pairing score for Y dataset features}
\item{adductX}{adduct label of features from dataset X}
\item{adductY}{adduct label of features from dataset Y}
\item{...}{Sample and extra columns from both datasets X & Y}
}
\description{
Returns the constructed table reporting every possible metabolomics
feature pair alignment, as determined from m/z grouping step.
}
