% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combiner_package_doc.R
\docType{package}
\name{Combiner}
\alias{Combiner}
\alias{Combiner-package}
\title{Combiner: Method for Combining LC-MS Metabolomics Feature Measurements}
\description{
Takes two untargeted metabolomics feature lists (consisting of m/z,
rt, and sample intensity measurements, plus optional identifiers & adduct
labels) and outputs a merged feature list consisting of potential
compound matches, ranked by a similarity score for groups of features. Inputs
are assumed to be derived from biologically similar samples analyzed with
a similar analytical method.
}
