% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinerError.R
\name{combinerError}
\alias{combinerError}
\title{Obtain Errors for Combiner Object Checks}
\usage{
combinerError(errNo, type)
}
\arguments{
\item{errNo}{integer error code.}

\item{type}{character object type (either "combinerTable", "metabCombiner"
or "metabData")}
}
\value{
A customized error message for specific object check.
}
\description{
This function stores and returns a customized error message
when checking the validity of certain objects.
}
