% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{crossValidationGAM}
\alias{crossValidationGAM}
\title{Cross Validation for GAM model}
\usage{
crossValidationGAM(rts, k, bs, m, family, method, optim, seed, ...)
}
\arguments{
\item{rts}{}

\item{k}{}

\item{bs}{}

\item{m}{}

\item{family}{}

\item{method}{}

\item{optim}{}

\item{seed}{integer. Psuedo-random seed generator for cross-validation.}
}
\value{
Optimal value for \code{k} as determined by 10-fold cross validation.
}
\description{
Helper function for \code{fit_gam()}. Determines optimal value
of \code{k} basis functions (among user-defined choices) for Generalized
Additive Model, using a 10-fold cross validation.
}
