% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectFields.R
\name{detectFields}
\alias{detectFields}
\title{Detect metabData Input Columns}
\usage{
detectFields(Data, table, mz, rt, id, adduct, samples, extra, Q)
}
\arguments{
\item{Data}{a metabData object.}

\item{table}{data frame containing metabolomics features or path to
metabolomics data file.}

\item{mz}{Character. Name for column containing m/z values. If "mz",
will search for {mz, m/z, m.z, mass}}

\item{rt}{Character. Name for column containing retention time values.
If "rt", will search for {rt, retention time, r.t, RT}.}

\item{id}{Character. Name for column containing compound identifiers.
If "id", will search for {id, compound, feature}}

\item{adduct}{Character. Name for column containing adduct labels. If
"adduct", will search for {adduct, adducts, annotation}.}

\item{samples}{Character. Names of columns containing sample values. If
"detect", finds longest stretch of consecutive numeric columns.}

\item{extra}{Character. Names of (additional) user-supplied columns.}

\item{Q}{Character. Name of user-supplied abundance quantiles column.
If null, quantiles will be automatically calculated.}
}
\value{
an initialized and formatted \code{metabData} object.
}
\description{
Detect metabData Input Columns
}
