% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-featdata.R
\name{featdata}
\alias{featdata}
\alias{featdata,metabCombiner-method}
\title{Obtain Feature Metadata}
\usage{
featdata(object, data = NULL)

\S4method{featdata}{metabCombiner}(object, data = NULL)
}
\arguments{
\item{object}{a \code{metabCombiner} object}

\item{data}{character dataset identifier}
}
\value{
data frame of feature metadata from one or all datasets
}
\description{
\code{metabCombiner} objects organize metabolomics feature
information in the "featdata" slot. This method retrieves all metadata
or that of one dataset. The rows should identically correspond to the same
rows from the combinedTable data frame.
}
\examples{
data(plasma30)
data(plasma20)

p30 <- metabData(head(plasma30,500), samples = "CHEAR")
p20 <- metabData(head(plasma20,500), samples = "Red")

p.comb <- metabCombiner(p30, p20, xid = "p30", yid = "p20")

#full metadata extraction
fdata <- featdata(p.comb, data = NULL)

#single dataset feature information extraction
fdata <- featdata(p.comb, data = "p20")

}
