% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_loess.R
\name{filterAnchorsLoess}
\alias{filterAnchorsLoess}
\title{Select Out High Residual Ordered Pairs}
\usage{
filterAnchorsLoess(rts, spans, iterFilter, ratio, frac, loess.parameters)
}
\description{
This is a helper function for the \code{fit_loess} function. Filters anchor retention
time ordered pairs using the residuals calculated from
}
\seealso{
[filterAnchorsGAM],[fit_loess], [crossValidationLoess]
}
