% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formCombinedTable.R
\name{formCombinedTable}
\alias{formCombinedTable}
\title{Form Combiner Report Table}
\usage{
formCombinedTable(object, xset, yset, nGroups)
}
\arguments{
\item{object}{metabCombiner object}

\item{xset}{data frame. A processed metabolomics feature table.}

\item{yset}{data frame. A processed metabolomics feature table.}

\item{nGroups}{integer. Total number of computed feature groups}
}
\value{
metabCombiner object with initialized combinedTable data frame.
}
\description{
Takes previously computed m/z groups using \code{mzGroup()}
and creates merged \code{combinedTable} consisting of all possible feature
alignments, all initialized with equivalent score and ranking.
}
