% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formCombinerTable.R
\name{formCombinerTable}
\alias{formCombinerTable}
\title{Form Combiner Report Table}
\usage{
formCombinerTable(object, xdata, ydata, nGroups)
}
\arguments{
\item{object}{metabCombiner object}

\item{xdata}{data frame. A processed metabolomics feature table.}

\item{ydata}{data frame. A processed metabolomics feature table.}

\item{nGroups}{integer. Total number of computed feature groups}
}
\value{
metabCombiner object with initialized combinerTable data frame.
}
\description{
Takes previously computed m/z groups using \code{mzGroup()}
and creates merged \code{combinerTable} consisting of all possible feature
alignments, all initialized with equivalent score and ranking.
}
