% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{getAnchors,metabCombiner-method}
\alias{getAnchors,metabCombiner-method}
\title{Get Ordered Retention Time Pairs}
\usage{
\S4method{getAnchors}{metabCombiner}(object)
}
\arguments{
\item{object}{metabCombiner object}
}
\value{
Data frame of anchor feature alignments.
}
\description{
This returns the list of feature alignments used to anchor a retention
time projection model, constructed by \code{\link{selectAnchors}}.
}
\seealso{
[selectAnchors]
}
