% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{getCoefficients,metabCombiner-method}
\alias{getCoefficients,metabCombiner-method}
\title{Obtain last-used set of score coefficients}
\usage{
\S4method{getCoefficients}{metabCombiner}(object)
}
\arguments{
\item{object}{metabCombiner object}
}
\value{
A list of the last used weight parameters:
\item{A}{Specific weight penalizing feature m/z differences}
\item{B}{Specific weight penalizing relative error of retention time projection}
\item{C}{Specific weight penalizing differences in abundance quantiles}
}
\description{
Provides the last used weight arguments from \code{calcScores()} function.
Returns NULL if \code{calcScores()} has not yet been called.
}
