% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabData.R
\docType{methods}
\name{getData,metabData-method}
\alias{getData,metabData-method}
\title{Obtain processed metabolomics data frame}
\usage{
\S4method{getData}{metabData}(object)
}
\arguments{
\item{object}{A metabData object}
}
\value{
Processed metabolomics feature data frame.
}
\description{
Extracts the formatted feature table contained within metabData object.
}
\examples{

}
