% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-metabData.R
\name{getData}
\alias{getData}
\alias{getData,metabData-method}
\title{Get Processed Dataset}
\usage{
getData(object)

\S4method{getData}{metabData}(object)
}
\arguments{
\item{object}{metabData object}
}
\value{
Single Metabolomics Data Frame
}
\description{
Get Processed Dataset
}
\examples{
data(plasma30)

p30 <- metabData(plasma30, samples = "CHEAR")
data <- getData(p30)

}
