% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabData.R
\docType{methods}
\name{getExtra,metabData-method}
\alias{getExtra,metabData-method}
\title{Get Names of Additional Data Fields}
\usage{
\S4method{getExtra}{metabData}(object)
}
\arguments{
\item{object}{A metabData object}
}
\value{
Names of 'extra' columns in \code{data} field of metabData
}
\description{
Get Names of Additional Data Fields
}
