% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-metabData.R
\docType{methods}
\name{getExtra}
\alias{getExtra}
\alias{getExtra,metabData-method}
\title{Get Extra Data Column Names}
\usage{
getExtra(object, data = c("x", "y"))

\S4method{getExtra}{metabData}(object)
}
\arguments{
\item{object}{metabCombiner or metabData object}

\item{data}{Choice of input dataset, 'x' or 'y'}
}
\value{
character vector of extra column names
}
\description{
Get Extra Data Column Names
}
\section{Methods (by class)}{
\itemize{
\item \code{metabData}: Method for "metabData" Objects
}}

