% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{getModel,metabCombiner-method}
\alias{getModel,metabCombiner-method}
\title{Get Fitted RT Model}
\usage{
\S4method{getModel}{metabCombiner}(object, fit = c("gam", "loess"))
}
\arguments{
\item{object}{metabCombiner object}

\item{fit}{Choice of model. Either "loess" or "gam".}
}
\value{
Nonlinear retention time fitting model.
}
\description{
This returns the last fitted RT projected model from a metabCombiner object
of type "gam" or "loess".
}
\seealso{
\code{\link{fit_gam}}, \code{\link{fit_loess}}
}
