% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-metabCombiner.R
\docType{methods}
\name{getModel}
\alias{getModel}
\alias{getModel,metabCombiner-method}
\title{Get Fitted RT Model}
\usage{
getModel(object, fit = c("gam", "loess"))

\S4method{getModel}{metabCombiner}(object, fit = c("gam", "loess"))
}
\arguments{
\item{object}{metabCombiner object}

\item{fit}{Choice of model, "gam" or "loess"}
}
\value{
nonlinear retention time fit object
}
\description{
Returns the last fitted RT projected model from a metabCombiner object of
type "gam" or "loess".
}
\section{Methods (by class)}{
\itemize{
\item \code{metabCombiner}: Method for 'metabCombiner' object
}}

\seealso{
\code{\link{fit_gam}}, \code{\link{fit_loess}}
}
