% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{getSamples,metabCombiner-method}
\alias{getSamples,metabCombiner-method}
\title{Get Sample Names From metabCombiner Object}
\usage{
\S4method{getSamples}{metabCombiner}(object, data = c("x", "y"))
}
\arguments{
\item{object}{metabCombiner object}

\item{data}{Either one of 'x' or 'y'.}
}
\value{
If data is "x", returns sample names from xdata; if "y", returns
sample names from ydata.
}
\description{
Get Sample Names From metabCombiner Object
}
