% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabData.R
\docType{methods}
\name{getSamples,metabData-method}
\alias{getSamples,metabData-method}
\title{Obtain Metabolomics Sample Names}
\usage{
\S4method{getSamples}{metabData}(object)
}
\arguments{
\item{object}{A metabData object}
}
\value{
Names of samples of formatted dataset.
}
\description{
Obtain Metabolomics Sample Names
}
