% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{getStats,metabCombiner-method}
\alias{getStats,metabCombiner-method}
\title{Get metabCombiner Statistics}
\usage{
\S4method{getStats}{metabCombiner}(object)
}
\arguments{
\item{object}{metabCombiner object}
}
\value{
A list containing statistics from metabCombiner object.
}
\description{
This function prints a set of metabCombiner object statistics.
}
