% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabData.R
\docType{methods}
\name{getStats,metabData-method}
\alias{getStats,metabData-method}
\title{Obtain metabData Statistics}
\usage{
\S4method{getStats}{metabData}(object)
}
\arguments{
\item{object}{A metabData object}
}
\value{
Feature statistics from metabData object
}
\description{
Obtain metabData Statistics
}
