% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-metabCombiner.R,
%   R/methods-metabData.R
\docType{methods}
\name{getStats}
\alias{getStats}
\alias{getStats,metabCombiner-method}
\alias{getStats,metabData-method}
\title{Get Object Statistics}
\usage{
getStats(object)

\S4method{getStats}{metabCombiner}(object)

\S4method{getStats}{metabData}(object)
}
\arguments{
\item{object}{metabCombiner or metabData object}
}
\value{
list of object-specific statistics
}
\description{
Get Object Statistics
}
\section{Methods (by class)}{
\itemize{
\item \code{metabCombiner}: Method for 'metabCombiner' object

\item \code{metabData}: Method for 'metabData' object
}}

