% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustData.R
\name{imputeVals}
\alias{imputeVals}
\title{Impute Missing Values for Metabolomics Features}
\usage{
imputeVals(data, samples, imputeVal)
}
\arguments{
\item{data}{constructed data frame of metabolomics features}

\item{samples}{names of experimental samples}

\item{imputeVal}{One of "median", "mean", or non-negative numeric value.}
}
\value{
Metabolomics data frame with imputed signal intensity values. If imputeVal is
equal to "median", feature missing values will be replaced with median.
If impute Val is equal to "mean", will impute the individual feature's
}
\description{
This is an optional method for imputing missing metabolomics feature
abundance values. This function is limited to imputation of feature
means, medians, or a constant value.
}
