% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinerError.R
\name{isCombinerTable}
\alias{isCombinerTable}
\title{Determine combinerTable Validity}
\usage{
isCombinerTable(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
0 if object is a valid Combiner Table; an integer code otherwise
}
\description{
Checks whether input object is a valid metabData.Returns an
integer code if invalid. Function is used alongside \code{combinerError}.
}
