% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinerCheck.R
\name{isMetabCombiner}
\alias{isMetabCombiner}
\title{Determine if object is a valid metabCombiner object}
\usage{
isMetabCombiner(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
0 if object is a valid metabData object; an integer code otherwise
}
\description{
Checks whether input object is a valid metabCombiner.Returns an
integer code if invalid. Function is used alongside \code{combinerCheck}.
}
