% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAnchors.R
\name{iterativeAnchorSelection}
\alias{iterativeAnchorSelection}
\title{Iteratively Select Ordered Feature Pairs}
\usage{
iterativeAnchorSelection(cTable, windX, windY, swap = FALSE)
}
\arguments{
\item{cTable}{data frame, contains only feature ids, mzs, rts, Qs, & labels}

\item{windX}{numeric positive retention time exclusion window in X dataset.}

\item{windY}{numeric positive retention time exclusion windown in Y dataset.}

\item{useID}{logical. Operation proceeds if TRUE, terminates early otherwise.}
}
\value{
Data frame
}
\description{

}
