% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{labelRowsParam}
\alias{labelRowsParam}
\title{List labelRows Defaults}
\usage{
labelRowsParam(
  maxRankX = 3,
  maxRankY = 3,
  minScore = 0.5,
  delta = 0.1,
  method = "score",
  balanced = TRUE,
  remove = FALSE,
  brackets_ignore = c("(", "[", "{")
)
}
\arguments{
\item{maxRankX}{maximum rank allowable for X features; default: 3}

\item{maxRankY}{maximum rank allowable for Y features; default: 3}

\item{minScore}{minimum score threshold; default: 0.5}

\item{delta}{score distance or mz/rt difference tolerances for subgrouping;
default: 0.1}

\item{method}{thresholding method for subgroup detection ("score" or "mzrt");
default: "score"}

\item{balanced}{option to reduce balanced groups; default: TRUE}

\item{remove}{option to eliminate rows determined as removable;
default: FALSE}

\item{brackets_ignore}{bracket types for ignoring string comparisons}
}
\value{
list of labelRows parameters
}
\description{
List of default parameters for combined table row annotation and removal.
See help(labelRows) or ?labelRows for more details.
}
\examples{
lrParams <- labelRowsParam(maxRankX = 2, maxRankY = 2, delta = 0.1)

}
\seealso{
\code{\link{labelRows}}, \code{\link{metabCombine}}
}
