% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabBatches}
\alias{metabBatches}
\title{Three LC-MS Metabolomics Batch Datasets}
\format{
A list containing three identically formatted data frames
}
\usage{
data(metabBatches)
}
\description{
An example multi-batch LC-MS metabolomics analysis of human
plasma, used to demonstrate \code{\link{batchCombine}}. Due to the large size
of the full experimental data, only three of the batches are loaded here with
a subset of the samples and features from each batch.
}
\keyword{datasets}
