% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabCombiner.R
\name{metabCombiner}
\alias{metabCombiner}
\title{Form a metabCombiner object.}
\usage{
metabCombiner(xdata, ydata, binGap = 0.005)
}
\arguments{
\item{xdata}{metabData object. One of two datasets to be Combined.}

\item{ydata}{metabData object. One of two datasets to be Combined.}

\item{binGap}{numeric. Parameter used for grouping features by m/z.
See ?mzGroup for more details.}
}
\value{
a metabCombiner object
}
\description{
\code{metabCombiner()} takes two metabolomics featurelists, contained in
\code{metabData} objects,and constructs a merged dataset containing groups of
features detected in both featurelists with similar m/z values.
}
\details{
The \code{binGap} argument defines consecutive differences between grouped
m/z features. \code{metabCombiner} objects are used for all subsequent
processing steps.
}
