% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metabData-class}
\alias{metabData-class}
\title{'metabData' Single Metabolomics Dataset Class}
\description{
This class is designed to process and format input metabolomics feature
tables. It stores the information from individual metabolomics datasets,
including the formatted feature table, sample names, and feature statistics.
}
\section{Slots}{

\describe{
\item{\code{data}}{formatted metabolomics data frame.}

\item{\code{samples}}{character vector of analyzed sample names}

\item{\code{extra}}{character vector of non-analyzed columns names}

\item{\code{stats}}{A list of dataset statistics}
}}

