% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabData.R
\name{metabData}
\alias{metabData}
\title{Constructor for the metabData object.}
\usage{
metabData(table, mz = "mz", rt = "rt", id = "id",
  adduct = "adduct", samples = NULL, Q = NULL, extra = NULL,
  rtmin = "min", rtmax = "max", misspc = 50, measure = c("median",
  "mean"), zero = FALSE, impute = FALSE, imputeVal = 100,
  duplicate = c(0.0025, 0.05))
}
\arguments{
\item{table}{Path to file containing feature table or data.frame object
containing features}

\item{mz}{Character. Name for column containing m/z values. If "mz",
will search for {mz, m/z, m.z, mass}}

\item{rt}{Character. Name for column containing retention time values.
If "rt", will search for {rt, retention time, r.t, RT}.}

\item{id}{Character. Name for column containing compound identifiers.
If "id", will search for {id, compound, feature}}

\item{adduct}{Character. Name for column containing adduct labels. If
"adduct", will search for {adduct, adducts, annotation}.}

\item{samples}{Character. Names of columns containing sample values. If
"detect", finds longest stretch of consecutive numeric columns.}

\item{Q}{Character. Name of user supplied quantile column.}

\item{extra}{Character. Names of (additional) user-supplied columns.}

\item{rtmin}{Numeric. Minimum retention time for analysis.}

\item{rtmax}{Numeric. Maximum retention time for analysis.}

\item{misspc}{Numeric. Threshold missingness percentage for analysis.}

\item{measure}{Character. Central abundance measure, either "median" or "mean".}

\item{zero}{Logical. Whether to consider zero values as missing.}

\item{impute}{Logical. Whether to impute value for missing sample measures.}

\item{imputeVal}{Imputed value. One of "median", "mean", or numeric value.}

\item{duplicate}{Numeric ordered pair (m/z, rt) duplicate feature tolerances.}
}
\value{
An object of class metabData containing the specific information
specified by \code{mz,rt, samples, id, adduct, Q, and extra} arguments, and adjusted
by pre-processing steps.
}
\description{
This is a constructor for objects of type \code{metabData}. A processed
  metabolomics feature data frame must contain columns for m/z, rt, and numeric
  sample intensities. The user may also supply some optional fields such as
  \code{id} and \code{adduct} label columns. Non-analyzed columns can be
  included into the final output by specifying the names of these columns in the
  "extra" argument. All required arguments are checked for correctness and
  validity (e.g. no negative m/z or rt values, no overlapping columns, etc...).

Following construction is a pre-analysis filtering of rows that are either,
  1) Outside of the specified retention time range (\code{rtmin},\code{rtmax}),
  2) Missing in excess of \code{misspc} percent of analyzed samples, or 3)
  deemed duplicates by small pairwise differences (mzTol,rtTol) as specified
  by the \code{duplicate} argument. Remaining features are ranked by abundance
  quantiles, Q, using a central \code{measure}, either "median" or "mean."
  Alternatively, the abundance quantiles column can be specified in the
  argument \code{Q}.

\code{metabData()} currently supports a limited set of options for missing value
  imputation, specified by the \code{impute} & \code{imputeVal} options.
}
