% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-metabCombiner.R
\docType{methods}
\name{nongrouped}
\alias{nongrouped}
\alias{nongrouped,metabCombiner-method}
\title{Get Nongrouped Features}
\usage{
nongrouped(object, data = c("x", "y"))

\S4method{nongrouped}{metabCombiner}(object, data = c("x", "y"))
}
\arguments{
\item{object}{metabCombiner object}

\item{data}{Either one of 'x' or 'y'}
}
\value{
If data is "x", returns non-grouped X features ; if "y", returns
        non-grouped Y features
}
\description{
Features that lack a close m/z counterpart in the complementary dataset are
not grouped and can be obtained from this method. The \code{data} argument
specifies from which dataset the non-grouped feature comes from.
}
\section{Methods (by class)}{
\itemize{
\item \code{metabCombiner}: Method for 'metabCombiner' objects
}}

