% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateParams.R
\name{objective}
\alias{objective}
\title{Weight Parameter Objective Function}
\usage{
objective(cTable, identities, A, B, C, minScore, mzdiff, rtdiff, qdiff,
  rtrange, adductdiff, penalty, matches, mismatches)
}
\arguments{
\item{cTable}{}

\item{identities}{}

\item{A}{}

\item{B}{}

\item{C}{}

\item{minScore}{}

\item{mzdiff}{}

\item{rtdiff}{}

\item{qdiff}{}

\item{rtrange}{}

\item{adductdiff}{}

\item{penalty}{}

\item{matches}{}

\item{mismatches}{}
}
\value{

}
\description{

}
