% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{plot,metabCombiner-method}
\alias{plot,metabCombiner-method}
\title{Plot nonlinear fit between retention times}
\usage{
\S4method{plot}{metabCombiner}(object, fit = c("gam", "loess"), pcol,
  lcol, lwd, ...)
}
\arguments{
\item{object}{metabCombiner object}

\item{fit}{Choice of model. Either "gam" or "loess".}

\item{pcol}{color of the points (ordered pairs) in the plot.}

\item{lcol}{color of the fitted line in the plot}

\item{...}{Other variables passed into graphics::plot}
}
\description{
Plot nonlinear fit between retention times
}
