% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Combiner.R
\name{plot_Combiner}
\alias{plot_Combiner}
\title{Plot Combiner Fits}
\usage{
plot_Combiner(object, fit = c("gam", "loess"), pcol, lcol, lwd, ...)
}
\arguments{
\item{object}{metabCombiner object}

\item{fit}{Choice of model (either "gam" or "loess").}

\item{pcol}{color of the points (ordered pairs) in the plot.}

\item{lcol}{color of the fitted line in the plot}

\item{lwd}{line width of the curve fit between anchor points.}

\item{...}{Other variables passed into graphics::plot}
}
\description{
This is the plotting method for Combiner objects. Displays anchoring ordered
pairs and a curve fit computed using \code{fit_gam} or \code{fit_loess}.
}
