% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceTable.R
\name{reduceTable}
\alias{reduceTable}
\title{Label and Remove Combiner Report Rows}
\usage{
reduceTable(object, maxRankX = 3, maxRankY = 3, minScore = 0.3,
  balanced = TRUE, conflict = c(0.003, 0.3, 0.003, 0.3),
  remove = FALSE)
}
\arguments{
\item{object}{Either a \code{metabCombiner} object or \code{combinerTable}.}

\item{maxRankX}{Integer. Maximum allowable rank for X dataset features.
Must be greater than 1.}

\item{maxRankY}{Integer. Maximum allowable rank for Y dataset features.
Must be greater than 1.}

\item{minScore}{Numeric. Minimum allowable score (between 0 & 1) for
metabolomics feature matches.}

\item{balanced}{Logical. Optional processing of "balanced" groups (defined as
groups with equal numbers of features from X & Y datasets).
If no contradicting top-matches,}

\item{conflict}{numeric vector. Non-negative Tolerance values for determining
if pairs of features within a single dataset are considered
"conflicting. Must be a length 4 vector. The first and second
values give the m/z and rt tolerances, respectively, for
dataset X; third and fourth value give the m/z tolerance and rt
tolerances, respectively, for dataset Y.}

\item{remove}{Logical. Option to keep or discard rows deemed removable (labeled
"REMOVE" in Combiner Report)}
}
\value{
combinerTable with labeled column for removable rows
}
\description{
Updates the \code{combinerTable} with labeled determinations of which rows are
removable, as well as conflicting and identity-labeled rows. Optionally, these
removable rows may be discarded, resulting in a reduced \code{combinerTable}.
}
\details{
Combiner initially reports all possible feature alignments in the rows of
\code{combinerTable} report. Most of these alignments are inaccurate and
require inspection and removal. This function is used to automate most of the
reduction process by labeling certain rows as removable or conflicting, based
on certain conditions, and is performed after

The labeling rules are as follows:
1) Rows with matching idx & idy strings are labeled "IDENTITY". These rows are
not labeled "REMOVE", irrespective of subsequent criteria.
2) Rows with a score < \code{minScore}: label "REMOVE"
3) Rows with rankX > \code{maxRankX} or rankY > \code{maxRankY}: label "REMOVE"
4) If row has score > \code{minScore} and 1 < rankX \eqn{\leq} \code{maxRankX}:
   or 1 < rankY \eqn{\leq}\code{maxRankY}:
}
