% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAnchors.R
\name{selectAnchors}
\alias{selectAnchors}
\title{Select Ordered Pair Features for Nonlinear RT Model}
\usage{
selectAnchors(object, useID = FALSE, tolMZ = 0.005, tolQ = 0.5,
  windX = 0.03, windY = 0.03)
}
\arguments{
\item{object}{metabCombiner object.}

\item{useID}{logical. Option to first search for IDs as anchors.}

\item{tolMZ}{numeric. m/z tolerance for prospective anchors.}

\item{tolQ}{numeric. Quantile Q tolerance for prospective anchors.}

\item{windX}{numeric. Retention time exclusion window around each anchor in
X dataset.}

\item{windY}{numeric. Retention time exclusion window around each anchor in
Y dataset.}
}
\value{
an updated metabCombiner object containing a table of ordered feature
pairs for fitting a nonlinear rt fitting model.
}
\description{
Select Ordered Pair Features for Nonlinear RT Model
}
\examples{


}
