% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustData.R
\name{adjustData}
\alias{adjustData}
\title{Process and Filter Metabolomics Feature Lists}
\usage{
adjustData(Data, misspc, measure, rtmin, rtmax, zero, duplicate)
}
\arguments{
\item{Data}{a metabData object.}

\item{misspc}{Numeric. Threshold missingness percentage for analysis.}

\item{measure}{Character. Central abundance measure, either "median" or "mean".}

\item{rtmin}{Numeric. Minimum retention time for analysis.}

\item{rtmax}{Numeric. Maximum retention time for analysis.}

\item{zero}{Logical. Whether to consider zero values as missing.}

\item{duplicate}{Ordered numeric pair (m/z, rt) tolerance parameters for
duplicate feature search.}
}
\value{
An updated metabData object. The \code{data} field is processed by the listed
steps and
}
\description{
\code{adjustData()} contains a set of pre-analysis steps for processing LC-MS
metabolomics feature tables.
}
\details{
The pre-analysis adjustment steps include:
1) Restriction to a feature retention time range \code{rtmin} \eqn{\le}
   rt \eqn{\le} \code{rtmax}
2) Removal of features with a missing value percentage exceeding \code{misspc}
3) Removal of duplicate metabolomics features.

After reducing the table, abundance quantile values are calculated for the
remaining features (unless initially supplied by the user). The \code{measure}
argument determines which central abundance statistic ("mean" or "median") is
used to rank features.
}
\seealso{
\code{\link{metabData}}: the constructor for metabData objects,
\code{\link{filterRT}}: function for filtering by retention times,
\code{\link{findDuplicates}}: function for finding duplicates
}
