% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcScores.R
\name{calcScores}
\alias{calcScores}
\title{Compute Feature Similarity Scores}
\usage{
calcScores(object, A, B, C, fit = c("gam", "loess"), usePPM = FALSE,
  useAdduct = FALSE, adduct = 2, groups = NULL)
}
\arguments{
\item{object}{metabCombiner object.}

\item{A}{Numeric weight for penalizing m/z differences.}

\item{B}{Numeric weight for penalizing differences between fitted & observed
retention times}

\item{C}{Numeric weight for penalizing differences in Q (abundance quantiles).}

\item{fit}{Character. Choice of fitted rt model, "gam" or "loess."}

\item{usePPM}{logical. Option to use relative (as opposed to absolute) m/z
differences in score computations.}

\item{useAdduct}{logical. Option to penalize mismatches in (non-empty) adduct
column labels.}

\item{adduct}{numeric. If useAdduct is TRUE, divides mismatching and non-empty
adduct column labels by this value.}

\item{groups}{integer. Vector of feature groups to score. If set to NULL
(default), will compute scores for all feature groups.}
}
\value{
metabCombiner object with updated combinerTable. rtProj column will
contain fitted retention times determined from previously computed model;
score will contain the computed pairwise similarity scores of features from
datasets x & y; rankX & rankY are the integer ranks of scores for x & y
features in descending order.
}
\description{
Calculates a pairwise similarity (between 0 & 1) between all
grouped features in \code{metabCombiner} object.
}
