% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinerError.R
\name{combinerError}
\alias{combinerError}
\title{Obtain Errors for Combiner Object Checks}
\usage{
combinerError(errNo, type)
}
\arguments{
\item{errNo}{integer error code.}

\item{type}{character object type (either "combinerTable", "metabCombiner"
or "metabData")}
}
\value{
A customized error message for specific object check.
}
\description{
This function stores and returns a customized error message
when checking the validity of certain objects.
}
\details{
In certain functions, an object must be checked for correctness. A
\code{metabData} must have a properly formatted dataset with the correct
column names & types.A \code{metabCombiner} must have properly formatted
\code{combinerTable}, with expected names and columns. If one of these
conditions is not met, a non-zero numeric code is returned and this function
is used to print a specific error message corresponding to the appropriate
object and error code.
}
