% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{crossValidationGAM}
\alias{crossValidationGAM}
\title{Cross Validation for GAM model}
\usage{
crossValidationGAM(rts, k, bs, m, family, method, optimizer, seed, ...)
}
\arguments{
\item{rts}{Data frame of ordered retention time pairs.}

\item{k}{Integer vector. Selection of values to control the number of basis
functions for gam construction.}

\item{bs}{character. Choice of spline method from mgcv, either "bs" or "ps"}

\item{m}{integer. Basis and penalty order for GAM; see ?mgcv::s}

\item{family}{character. Choice of mgcv family; see: ?mgcv::family.mgcv}

\item{method}{character. Smoothing parameter estimation method; see:
?mgcv::gam}

\item{optimizer}{character. Method to optimize smoothing parameter; see:
?mgcv::gam}

\item{seed}{integer. Psuedo-random seed generator for cross-validation.}

\item{...}{Other arguments passed to \code{mgcv::gam}.}
}
\value{
Optimal value for \code{k} as determined by 10-fold cross validation.
}
\description{
Helper function for \code{fit_gam()}. Determines optimal value
of \code{k} basis functions (among user-defined choices) for Generalized
Additive Model, using a 10-fold cross validation.
}
