% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_loess.R
\name{crossValidationLoess}
\alias{crossValidationLoess}
\title{Cross Validation for Loess Model}
\usage{
crossValidationLoess(rts, spans, seed, loess.parameters)
}
\arguments{
\item{rts}{Data frame of ordered retention time pairs.}

\item{spans}{numeric vector. Selection of values (between 0 & 1) to control
the degree of smoothing of loess curve.}

\item{seed}{integer. Psuedo-random seed generator for cross validation.}

\item{loess.parameters}{Parameters for LOESS fitting. See ?loess.control.}
}
\value{
Optimal value among pre-selected values for \code{spans}.
}
\description{
Cross Validation for Loess Model
}
