% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{filterAnchors}
\alias{filterAnchors}
\title{Filter Outlier Ordered Pairs}
\usage{
filterAnchors(
  rts,
  fit,
  vals,
  iterFilter,
  ratio,
  frac,
  bs,
  m,
  family,
  method,
  optimizer,
  loess.pars,
  ...
)
}
\arguments{
\item{rts}{Data frame of ordered retention time pairs.}

\item{fit}{Either "gam" for GAM fits, or "loess" for loess fits}

\item{vals}{numeric values: k values for GAM fits, spans for loess fits}

\item{iterFilter}{integer number of residual filtering iterations}

\item{ratio}{numeric. A point is an outlier if the ratio of residual to
mean residual of a fit exceeds this value. Must be greater than 1.}

\item{frac}{numeric. A point is excluded if deemed a residual in more than
this fraction value times the number of fits. Must be between 0 & 1.}

\item{bs}{character. Choice of spline method from mgcv; either "bs" or "ps"}

\item{m}{integer. Basis and penalty order for GAM; see ?mgcv::s}

\item{family}{character. Choice of mgcv family; see: ?mgcv::family.mgcv}

\item{method}{character. Smoothing parameter estimation method; see:
?mgcv::gam}

\item{optimizer}{character. Method to optimize smoothing parameter; see:
?mgcv::gam}

\item{loess.pars}{parameters for LOESS fitting; see ?loess.control}

\item{...}{other arguments passed to \code{mgcv::gam}.}
}
\value{
anchor rts data frame with updated weights.
}
\description{
Helper function for \code{\link{fit_gam}} & \code{\link{fit_loess}}. It
filters the set of ordered pairs using the residuals calculated from
multiple GAM / loess fits.
}
