% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{filterAnchorsGAM}
\alias{filterAnchorsGAM}
\title{Filter Outlier Ordered Pairs}
\usage{
filterAnchorsGAM(rts, k, bs, m, family, method, iterFilter, optim, ratio,
  frac, ...)
}
\arguments{
\item{rts}{}

\item{k}{}

\item{bs}{}

\item{m}{}

\item{family}{}

\item{method}{}

\item{iterFilter}{}

\item{optim}{}

\item{ratio}{}

\item{frac}{}
}
\description{
Helper function for fit_gam.
}
