% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_loess.R
\name{filterAnchorsLoess}
\alias{filterAnchorsLoess}
\title{Filter Outlier Ordered Pairs}
\usage{
filterAnchorsLoess(rts, spans, iterFilter, ratio, frac, loess.parameters)
}
\arguments{
\item{rts}{Data frame of ordered retention time pairs.}

\item{spans}{numeric vector. Selection of values (between 0 & 1) to control
the degree of smoothing of loess curve.}

\item{iterFilter}{integer. Number of residual filtering iterations to perform.}

\item{ratio}{numeric. A point is considered an outlier if ratio of residual to
mean residual of a fit exceeds this value. Must be greater than 1.}

\item{frac}{numeric. A point is excluded if deemed a residual in more than
this fraction value times the number of fits. Must be between 0 & 1.}

\item{loess.parameters}{Parameters for LOESS fitting. See ?loess.control.}
}
\value{
Ordered retention time pairs data frame with updated weights.
}
\description{
This is a helper function for \code{fit_loess}. It filters retention time
ordered pairs using the residuals calculated from multiple LOESS fits.
}
\seealso{
[filterAnchorsGAM],[fit_loess], [crossValidationLoess]
}
