% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit GAM model between retention times}
\usage{
fit_gam(object, useID = FALSE, k = seq(10, 25, by = 5),
  iterFilter = 2, ratio = 2, frac = 0.5, bs = c("bs", "ps", "tp"),
  family = c("scat", "gaussian"), weights = 1, m = c(3, 2),
  method = "REML", optim = "newton", seed = 100, ...)
}
\arguments{
\item{object}{a metabCombiner object.}

\item{useID}{logical. Option to use matched IDs to inform modeling process.}

\item{k}{integer vector. Selection of values to control the number of basis
functions for gam construction. Best value chosen by 10-fold cross validation.}

\item{iterFilter}{integer. Number of residual filtering iterations to perform.}

\item{ratio}{numeric. Residual multiplier for determining outliers.}

\item{bs}{character. Choice of spline method from mgcv; see: ?smooth.terms}

\item{family}{character. Choice of mgcv family; see: ?family.mgcv}

\item{m}{integer. Basis and penalty order for GAM; see ?mgcv::s}

\item{method}{character. Smoothing parameter estimation method; see: ?gam}

\item{seed}{integer. Psuedo-random seed generator for cross validation.}

\item{optimizer}{character. Method to optimize smoothing parameter; see: ?gam}
}
\value{
Updated metabCombiner object with a GAM model
}
\description{
Fit a Generalized Additive Model between selected ordered retention 
time pairs
}
