% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_loess.R
\name{fit_loess}
\alias{fit_loess}
\title{Fit a LOESS Through Retention Time Ordered Pairs}
\usage{
fit_loess(object, useID = FALSE, spans = seq(0.2, 0.4, by = 0.05),
  iterFilter = 2, ratio = 2, frac = 0.5, iterLoess = 10,
  weights = 1, seed = 100)
}
\arguments{
\item{object}{metabCombiner object.}

\item{useID}{logical. Option to not filter identity-labeled ordered pair.}

\item{spans}{numeric vector. Span values (between 0 & 1) used for loess fits}

\item{iterFilter}{integer. Number of residual filtering iterations to perform.}

\item{ratio}{numeric residual multiplier for determining outliers.}

\item{frac}{numeric. Threshold proportion of fits needed to determine if
ordered pair is an outlier.}

\item{iterLoess}{integer. Number of robustness iterations to perform in
\code{loess()}.See ?loess.control for more details.}

\item{weights}{numeric. Optional user supplied weights. Note: vector length
must be #num_anchors + 2 (first and last are endpoints).}

\item{seed}{integer. Psuedo-random seed generator for cross validation.}
}
\description{

}
