% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{getAnchors,metabCombiner-method}
\alias{getAnchors,metabCombiner-method}
\title{Get Ordered Retention Time Pairs}
\usage{
\S4method{getAnchors}{metabCombiner}(object)
}
\arguments{
\item{object}{metabCombiner object}
}
\value{
anchor set constructed using selectAnchors().
}
\description{
Get Ordered Retention Time Pairs
}
