% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getAnchors}
\alias{getAnchors}
\title{Get Ordered Retention Time Pairs}
\usage{
getAnchors(object)
}
\arguments{
\item{object}{metabCombiner object}
}
\value{
Data frame of anchor features
}
\description{
This returns the data frame of feature alignments used to anchor a retention
time projection model, constructed by \code{\link{selectAnchors}}.
}
\seealso{
\code{\link{selectAnchors}}
}
