% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Obtain Last-Used Score Coefficients}
\usage{
getCoefficients(object)
}
\arguments{
\item{object}{metabCombiner object}
}
\value{
A list of the last used weight parameters:
\item{A}{Specific weight penalizing feature m/z differences}
\item{B}{Specific weight penalizing relative error of retention time projection}
\item{C}{Specific weight penalizing differences in abundance quantiles}
}
\description{
Provides the last used weight arguments from \code{calcScores()} function.
Returns empty list if \code{calcScores()} has not yet been called.
}
