% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{getData,metabCombiner-method}
\alias{getData,metabCombiner-method}
\title{Get Datasets From metabCombiner Object.}
\usage{
\S4method{getData}{metabCombiner}(object, data = c("x", "y"))
}
\arguments{
\item{object}{metabCombiner object}

\item{data}{Either one of 'x' or 'y'.}
}
\value{
If data is "x", returns dataset X ; if "y", returns dataset Y.
}
\description{
\code{metabCombiner} objects consist of two formatted metabolomics feature
tables. This function allows for returning one of the two processed datasets
contained within the object.
}
