% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-metabData.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,metabData-method}
\title{Get Processed Dataset}
\usage{
getData(object)

\S4method{getData}{metabData}(object)
}
\arguments{
\item{object}{metabData object}
}
\value{
Single Metabolomics Data Frame
}
\description{
Get Processed Dataset
}
\section{Methods (by class)}{
\itemize{
\item \code{metabData}: Method for "metabData" objects
}}

