% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-metabCombiner.R,
%   R/methods-metabData.R
\name{getExtra}
\alias{getExtra}
\alias{getExtra,metabCombiner-method}
\alias{getExtra,metabData-method}
\title{Get Extra Data Column Names}
\usage{
getExtra(object, data = "x")

\S4method{getExtra}{metabCombiner}(object, data = "x")

\S4method{getExtra}{metabData}(object)
}
\arguments{
\item{object}{metabCombiner or metabData object}

\item{data}{Choice of input dataset, e.g. "x" or "y"}
}
\value{
character vector of extra column names
}
\description{
Get Extra Data Column Names
}
\examples{
data(plasma30)
p30 <- metabData(plasma30, samples = "CHEAR", extra = "Red")
getExtra(p30)

}
