% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{getModel,metabCombiner-method}
\alias{getModel,metabCombiner-method}
\title{Get Fitted RT Model}
\usage{
\S4method{getModel}{metabCombiner}(object, fit = c("gam", "loess"))
}
\arguments{
\item{object}{\code{metabCombiner} object}

\item{fit}{Choice of either "gam" or "loess" model}
}
\description{
Returns the last fited RT projected model from a metabCombiner
object of type "gam" or "loess".
}
