% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getModel}
\alias{getModel}
\title{Get Fitted RT Model}
\usage{
getModel(object, fit = c("gam", "loess"))
}
\arguments{
\item{object}{metabCombiner object}

\item{fit}{Choice of model, "gam" or "loess"}
}
\value{
nonlinear retention time fit object
}
\description{
Returns the last fitted RT projected model from a metabCombiner object of
type "gam" or "loess".
}
\seealso{
\code{\link{fit_gam}}, \code{\link{fit_loess}}
}
