% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R
\docType{methods}
\name{getSamples,metabCombiner-method}
\alias{getSamples,metabCombiner-method}
\title{Get Sample Names From metabCombiner Object

\code{metabCombiner} objects consist of two formatted metabolomics feature
tables. This method returns the sample names from one of the two datasets.}
\usage{
\S4method{getSamples}{metabCombiner}(object, data = c("x", "y"))
}
\arguments{
\item{object}{metabCombiner object}

\item{data}{Character. One of either 'x' or 'y'.}
}
\value{
If data is "x", returns sample names for dataset X; if "y", returns
sample names from dataset Y.
}
\description{
Get Sample Names From metabCombiner Object

\code{metabCombiner} objects consist of two formatted metabolomics feature
tables. This method returns the sample names from one of the two datasets.
}
