% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getSamples}
\alias{getSamples}
\title{Get Sample Names From metabCombiner Object}
\usage{
getSamples(object, data = c("x", "y"))
}
\arguments{
\item{object}{metabCombiner or metabData object}

\item{data}{Character. One of either 'x' or 'y'.}
}
\value{
If data is "x", returns sample names for dataset X; if "y", returns
sample names from dataset Y.
}
\description{
\code{metabCombiner} objects consist of two formatted metabolomics
feature tables. This method returns the sample names from one of the two datasets.
}
