% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAnchors.R
\name{identityAnchorSelection}
\alias{identityAnchorSelection}
\title{Select Matching Ids as Anchors}
\usage{
identityAnchorSelection(cTable, windX, windY, useID)
}
\arguments{
\item{cTable}{data frame, contains only feature ids, mzs, rts, Qs, & labels}

\item{windX}{numeric positive retention time exclusion window in X dataset.}

\item{windY}{numeric positive retention time exclusion windown in Y dataset.}

\item{useID}{logical. Operation proceeds if TRUE, terminates otherwise.}
}
\description{
This is an optional helper function for \code{selectAnchors}. Uses identities
to guide selection of ordered retention time pairs. If useID option is set to
true, it will select pairs of features with matching ID character strings before
proceeding with iterative selection.
}
\note{
Identity anchors may violate constraints of m/z and Q difference
tolerances, and will not be removed if they fall within the exclusion window
of other features. If a name appears more than once, only one pair will be
selected, and the others excluded.
}
\seealso{
[selectAnchors], [selectIterativeAnchors]
}
