% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metabCombiner-class}
\alias{metabCombiner-class}
\title{'metabCombiner' Combined Metabolomics Dataset Class}
\description{
This is the main object for the \code{metabCombiner} package workflow. This
object holds a combined feature table, along with a retention time
projection model, the ordered pair anchors used to generate this model, and
key object statistics.
}
\section{Slots}{

\describe{
\item{\code{combinedTable}}{A data.frame combining tables measurements and displaying
all feature alignments}

\item{\code{nongrouped}}{A list of data frames containing features from input tables
that could not be paired in \code{combinedTable}}

\item{\code{anchors}}{A data.frame consisting of feature alignments to be used for
fitting a nonlinear retention time projection model}

\item{\code{model}}{A list containing the last fitted nonlinear model(s).}

\item{\code{coefficients}}{A list of last used A,B,C similarity weight values}

\item{\code{samples}}{A list of sample name character vectors from input datasets}

\item{\code{extra}}{A list of extra column chacter name vectors from input datasets}

\item{\code{stats}}{A set of useful metabCombiner statistics}
}}

