% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metabCombiner-class}
\alias{metabCombiner-class}
\title{'metabCombiner' Combined Metabolomics Dataset Class}
\description{
This is the main object for the \code{Combiner} package workflow. This
object holds a pair of analyzed metabolomics datasets and a combined
feature table, along with a retention time projection model.
}
\section{Slots}{

\describe{
\item{\code{xdata}}{One of two formatted metabData objects to be combined.}

\item{\code{ydata}}{One of two formatted metabData objects to be combined.}

\item{\code{combinerTable}}{A data.frame combining tables measurements and displaying
all feature alignments}

\item{\code{binGap}}{The numeric m/z gap used to compute feature groups}

\item{\code{anchors}}{A data.frame consisting of feature alignments to be used for
fitting a nonlinear retention time projection model}

\item{\code{model}}{A list containing the last fitted nonlinear model(s).}

\item{\code{coefficients}}{A list of last used A,B,C similarity score weight arguments}

\item{\code{stats}}{A list of useful metabCombiner object statistics}
}}

