% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabCombiner.R, R/metabCombiner_package_doc.R
\docType{package}
\name{metabCombiner}
\alias{metabCombiner}
\title{Form a metabCombiner object.}
\usage{
metabCombiner(xdata, ydata, binGap = 0.005)
}
\arguments{
\item{xdata}{metabData object. One of two datasets to be combined.}

\item{ydata}{metabData object. One of two datasets to be combined.}

\item{binGap}{numeric. Parameter used for grouping features by m/z.
See ?mzGroup for more details.}
}
\value{
a metabCombiner object
}
\description{
\code{metabCombiner()} takes two metabolomics featurelists, contained in
\code{metabData} objects and constructs a merged dataset containing groups
of features detected in both featurelists with similar m/z values.

Takes two untargeted metabolomics feature lists (consisting of m/z,
rt, and sample intensity measurements, plus optional identifiers & adduct
labels) and outputs a merged feature list consisting of potential
compound matches, ranked by a similarity score for groups of features. Inputs
are assumed to be derived from biologically similar samples analyzed with
a similar analytical method.
}
\details{
The \code{binGap} argument defines consecutive differences between grouped
m/z features. \code{metabCombiner} objects are used for all subsequent
processing steps.
}
\examples{
data(plasma30)
data(plasma20)

p30 <- metabData(plasma30, samples = "CHEAR")
p20 <- metabData(plasma20, samples = "Red", rtmax = 17.25)

p.comb = metabCombiner(xdata = p30, ydata = p20, binGap = 0.0075)

}
