% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metabData-class}
\alias{metabData-class}
\title{'metabData' Single Metabolomics Dataset Class}
\description{
This class is designed to pre-process, format, and facilitate the subsequent
combining of input metabolomics feature tables. It stores the information of
a single metabolomics dataset, including the formatted feature table, sample
names, and feature statistics.
}
\section{Slots}{

\describe{
\item{\code{data}}{A formatted metabolomics data frame.}

\item{\code{samples}}{A character vector of analyzed samples.}

\item{\code{extra}}{A character vector of names pertaining to non-analyzed columns}

\item{\code{stats}}{A list of dataset statistics}
}}

