% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabData.R
\name{metabData}
\alias{metabData}
\title{Constructor for the metabData object.}
\usage{
metabData(table, mz = "mz", rt = "rt", id = "id",
  adduct = "adduct", samples = NULL, Q = NULL, extra = NULL,
  rtmin = "min", rtmax = "max", misspc = 50, measure = c("median",
  "mean"), zero = FALSE, duplicate = c(0.0025, 0.05))
}
\arguments{
\item{table}{Path to file containing feature table or data.frame object
containing features}

\item{mz}{Character name(s) or regular expression associated with data
column containing m/z values. The first column whose name contains this
expression will be selected for analysis.}

\item{rt}{Character name(s) or regular expression associated with data
column containing retention time values. The first column whose name contains
this expression will be selected for analysis.}

\item{id}{Character name(s) or regular expression associated with data
column containing metabolomics feature identifiers. The first column whose
name contains this expression will be selected for analysis.}

\item{adduct}{Character name(s) or regular expression associated with data
column containing adduct, formula, or additional annotations. The first
column whose name contains this expression will be selected for analysis.}

\item{samples}{Character names of columns containing sample values. All
numeric columns containing these keywords are selected for analysis. If no
keywords given, will search for longest stretch of numeric columns remaining.}

\item{Q}{Character name(s) or regular expression associated with numeric
feature abundance quantiles. If NULL, abundance quantiles will be calculated
from sample intensities.}

\item{extra}{Character names of columns containing additional feature
information, e.g.  non-analyzed sample values. All columns containing these
keywords are selected for analysis.}

\item{rtmin}{Numeric. Minimum retention time for analysis.}

\item{rtmax}{Numeric. Maximum retention time for analysis.}

\item{misspc}{Numeric. Threshold missingness percentage for analysis.}

\item{measure}{Character. Central abundance measure, either "median" or "mean".}

\item{zero}{Logical. Whether to consider zero values as missing.}

\item{duplicate}{Numeric ordered pair (m/z, rt) duplicate feature tolerances.}
}
\value{
An object of class metabData containing the specific information
specified by \code{mz,rt, samples, id, adduct, Q, and extra} arguments, and
adjusted by pre-processing steps.
}
\description{
This is a constructor for objects of type \code{metabData}.
}
\details{
A processed metabolomics feature data frame must contain columns for m/z, rt,
and numeric sample intensities. The user may also supply some optional fields
such as identity (id) and adduct label columns. Non-analyzed columns can be
included into the final output by specifying the names of these columns in the
\code{extra} argument. All required arguments are checked for correctness,
validity (e.g. no negative m/z or rt values, each column is used at most once,
column types are valid, etc...),

Following construction is a pre-analysis filtering of rows that are either,
1) Outside of the specified retention time range (\code{rtmin},\code{rtmax}),
2) Missing in excess of \code{misspc} percent of analyzed samples, or
3) deemed duplicates by small pairwise differences (mzTol,rtTol) as specified
by the \code{duplicate} argument. Remaining features are ranked by abundance
quantiles, Q, using a central \code{measure}, either "median" or "mean."
Alternatively, the abundance quantiles column can be specified in the
argument \code{Q}.
}
