% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzGroup.R
\name{mzGroup}
\alias{mzGroup}
\title{Binning of mass spectral features in m/z dimension}
\usage{
mzGroup(object, binGap = 0.005)
}
\arguments{
\item{object}{metabCombiner object}

\item{binGap}{numeric. Gap between concecutive sorted pooled feature m/z values.}
}
\value{
metabCombiner object with group information in data objects
}
\description{
Features in two input feature lists are grouped by their m/z values.
}
\details{
The m/z values from both datasets are pooled, sorted, and binned by the
\code{binGap} argument. Feature groups form when there is at least one pair
of features from both datasets whose consecutive difference is less than
\code{binGap}.
}
