% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-metabCombiner.R
\name{nonmatched}
\alias{nonmatched}
\alias{nonmatched,metabCombiner-method}
\title{Get Nonmatched Features}
\usage{
nonmatched(object, data = c("x", "y"))

\S4method{nonmatched}{metabCombiner}(object, data = c("x", "y"))
}
\arguments{
\item{object}{metabCombiner object}

\item{data}{Either one of 'x' or 'y', specifying which dataset's nonmatched
features to return.}
}
\value{
If data is "x", returns non-matched X features ; if "y", returns
        non-matched Y features
}
\description{
Features that lack a any counterparts in the complementary dataset may be
obtained from this method.
}
\examples{
data(plasma30)
data(plasma20)

p30 <- metabData(plasma30, samples = "CHEAR")
p20 <- metabData(plasma20, samples = "Red", rtmax = 17.25)
p.comb = metabCombiner(xdata = p30, ydata = p20, binGap = 0.005)

nnmx = nonmatched(p.comb, data = "x")
nnmy = nonmatched(p.comb, data = "y")

}
