% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plasma20}
\alias{plasma20}
\title{20 minute Metabolomics Analysis of Human Plasma}
\format{
A data frame with 8910 rows and 22 columns:
\describe{
  \item{feature}{generic feature identifiers}
  \item{identity}{identities of feature (if known); feature identifier (if
                   unknown); names in parantheses if ambiguous)}
  \item{adduct}{adduct, fragment, or multimer annotations; labels in
                parantheses if ambiguous}
  \item{mz}{m/z values of metabolomics features}
  \item{rt}{rt values of metabolomics features}
  \item{...}{column names for CHEAR, Red Cross, Pooled, and Blank samples}
}
}
\usage{
data(plasma20)
}
\description{
An example metabolomics dataset consisting of human plasma from
Red Cross and CHEAR cohorts, plus pooled aliquots and blanks, acquired
with a 20 minute total Reversed-Phase Liquid Chromatography &
QTOF-MS instrument in the positive ionization mode.
}
\keyword{datasets}
