% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plasma30}
\alias{plasma30}
\title{30 minute Metabolomics Analysis of Human Plasma

An example metabolomics dataset consisting of human plasma from Red Cross
and CHEAR cohorts, plus pooled aliquots and negative controls, acquired
with a 30 minute total Reversed-Phase Liquid Chromatography and a QTOF
instrument in the positive ionization mode.

It is data frame with 8286 rows and 22 variables, including:
\itemize{
  \item{feature}{generic feature identifiers}
  \item{identity}{identities of feature (if known); feature identifier (if
                   unknown); names in parantheses if ambiguous)}
  \item{adduct}{adduct, fragment, or multimer annotations; labels in
                parantheses if ambiguous}
  \item{mz}{m/z values of metabolomics features}
  \item{rt}{rt values of metabolomics features}
  \item{...}{column names for CHEAR, Red Cross, Pooled, and Blank samples}
}}
\format{An object of class \code{data.frame} with 8286 rows and 22 columns.}
\usage{
plasma30
}
\description{
30 minute Metabolomics Analysis of Human Plasma

An example metabolomics dataset consisting of human plasma from Red Cross
and CHEAR cohorts, plus pooled aliquots and negative controls, acquired
with a 30 minute total Reversed-Phase Liquid Chromatography and a QTOF
instrument in the positive ionization mode.

It is data frame with 8286 rows and 22 variables, including:
\itemize{
  \item{feature}{generic feature identifiers}
  \item{identity}{identities of feature (if known); feature identifier (if
                   unknown); names in parantheses if ambiguous)}
  \item{adduct}{adduct, fragment, or multimer annotations; labels in
                parantheses if ambiguous}
  \item{mz}{m/z values of metabolomics features}
  \item{rt}{rt values of metabolomics features}
  \item{...}{column names for CHEAR, Red Cross, Pooled, and Blank samples}
}
}
\keyword{datasets}
