% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R, R/plot_Combiner.R
\docType{methods}
\name{plot,metabCombiner,ANY-method}
\alias{plot,metabCombiner,ANY-method}
\alias{plot_Combiner}
\title{Plot Combiner Fits}
\usage{
\S4method{plot}{metabCombiner,ANY}(x, y, ...)

plot_Combiner(x, fit = c("gam", "loess"), pcol, lcol, lwd, ...)
}
\arguments{
\item{x}{metabCombiner object}

\item{...}{Other variables passed into graphics::plot}

\item{fit}{Choice of model (either "gam" or "loess").}

\item{pcol}{color of the points (ordered pairs) in the plot.}

\item{lcol}{color of the fitted line in the plot}

\item{lwd}{line width of the curve fit between anchor points.}
}
\description{
This is the plotting method for Combiner objects. Displays anchoring ordered
pairs and a curve fit computed using \code{fit_gam} or \code{fit_loess}.
}
