% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metabCombiner.R, R/plot_Combiner.R
\docType{methods}
\name{plot,metabCombiner,ANY-method}
\alias{plot,metabCombiner,ANY-method}
\alias{plot_Combiner}
\title{Plot metabCombiner Fits}
\usage{
\S4method{plot}{metabCombiner,ANY}(x, y, ...)

plot_Combiner(object, fit = c("gam", "loess"), pcol, lcol, lwd, ...)
}
\arguments{
\item{x}{\code{metabCombiner} object}

\item{y}{...}

\item{...}{Other variables passed into graphics::plot}

\item{object}{metabCombiner object}

\item{fit}{Choice of model (either "gam" or "loess").}

\item{pcol}{color of the points (ordered pairs) in the plot.}

\item{lcol}{color of the fitted line in the plot}

\item{lwd}{line width of the curve fit between anchor points.}
}
\description{
This is a plotting method for metabCombiner objects. It displays anchoring
ordered pairs and a curve fit computed using \code{fit_gam} or
\code{fit_loess}, using R's basic graphics.
}
\examples{
\dontrun{
library(metabCombiner)
data(plasma30)
data(plasma20)

p30 <- metabData(plasma30, samples = "CHEAR")
p20 <- metabData(plasma20, samples = "Red", rtmax = 17.25)
p.combined = metabCombiner(xdata = p30, ydata = p20, binGap = 0.0075)
p.combined = selectAnchors(p.combined, tolMZ = 0.003, tolQ = 0.3, windY = 0.02)
p.combined = fit_gam(p.combined, k = seq(12,20,2), iterFilter = 1)
p.combined = fit_loess(p.combined, spans = seq(0.2,0.3, 0.02), iterFilter = 2)

##plot of GAM fit
plot(p.combined, main = "Example GAM Fit Plot", xlab = "X Dataset RTs",
     ylab = "Y Dataset RTs", pcol = "red", lcol = "blue", lwd = 5, pch = 19)
grid(lwd =  2, lty = 3 ) #adding gridlines

##plot of loess fit
plot(p.combined, fit = "loess", main = "Example loess Fit Plot",
     xlab = "X Dataset RTs", ylab = "Y Dataset RTs", pcol = "dark orange",
     lcol = "dark green", pch = 8)
}
}
