% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcScores.R
\name{scorePairs}
\alias{scorePairs}
\title{Calculate the pairwise score between grouped features}
\usage{
scorePairs(A, B, C, mzdiff, rtdiff, qdiff, rtrange, adductdiff)
}
\arguments{
\item{A}{Numeric weight for penalizing m/z differences.}

\item{B}{Numeric weight for penalizing differences between fitted & observed
retention times}

\item{C}{Numeric weight for differences in Q (abundance quantiles).}

\item{mzdiff}{Difference between feature m/z values; can be absolute or
relative (ppm)}

\item{rtdiff}{Difference between model-projected retention time value ('xdata')
& observed retention time ('ydata')}

\item{qdiff}{Difference between feature quantile Q values}

\item{adduct}{Numeric (x>=1). Divides computed score when adduct labels do
not match.}
}
\value{
Pairwise score between two grouped features(between 0 & 1) evaluating
         the likelihood of a compound match.
}
\description{
Helper function for \code{calcScores()} & \code{evalParams()}.
Calculates the pairwise similarity score between grouped features acquired
in complementary datasets.
}
\details{
Score between two grouped features x & y calculated as:

\deqn{S(x,y) = -exp(-A |mzx-mzy|- B |rty-rtproj|/rtrange - C |Qx-Qy|)}

where mzx & Qx correspond to the m/z and abundance quantile values of feature
x, mzy, rty, and Qy correspond to the m/z, retention time, and abundance
quantile values of feature y,  rtproj is the model-projected retention of
feature x onto the retention times of the dataset containing y, and rtrange is
the range of retention times of y dataset features. A, B, C are non-negative
constant weight parameters for penalizing m/z, rt, and Q  differences. Values
vary between 0 (no confidence combination) and 1 (high confidence combination).
}
