% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAnchors.R
\name{selectAnchors}
\alias{selectAnchors}
\title{Select Anchors for Nonlinear RT Model}
\usage{
selectAnchors(object, useID = FALSE, tolMZ = 0.005, tolQ = 0.5,
  windX = 0.03, windY = 0.03)
}
\arguments{
\item{object}{metabCombiner object.}

\item{useID}{logical. Option to first search for IDs as anchors.}

\item{tolMZ}{numeric. m/z tolerance for prospective anchors.}

\item{tolQ}{numeric. Quantile Q tolerance for prospective anchors.}

\item{windX}{numeric. Retention time exclusion window around each anchor in
X dataset. Optimal values are between 0.01 and 0.05 min (1-3s).}

\item{windY}{numeric. Retention time exclusion window around each anchor in
dataset Y. Optimal values are between 0.01 and 0.05 min (1-3s).}
}
\value{
an updated metabCombiner object containing a table of ordered feature
pairs for fitting a nonlinear rt fitting model.
}
\description{
This function selects a subset of possible alignments for use as ordered
pairs to anchor a retention time projection model. Alignments of abundant
features are prominent targets for anchor selection.
}
