% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAnchors.R
\name{selectAnchors}
\alias{selectAnchors}
\title{Select Anchors for Nonlinear RT Model}
\usage{
selectAnchors(object, useID = FALSE, tolMZ = 0.003, tolQ = 0.3,
  windX = 0.03, windY = 0.03, brackets_ignore = c("(", "[", "{"))
}
\arguments{
\item{object}{metabCombiner object.}

\item{useID}{logical. Option to first search for IDs as anchors.}

\item{tolMZ}{numeric. m/z tolerance for prospective anchors}

\item{tolQ}{numeric. Quantile Q tolerance for prospective anchors}

\item{windX}{numeric. Retention time exclusion window around each anchor in
X dataset. Optimal values are between 0.01 and 0.05 min (1-3s)}

\item{windY}{numeric. Retention time exclusion window around each anchor in
dataset Y. Optimal values are between 0.01 and 0.05 min (1-3s)}

\item{brackets_ignore}{If useID = TRUE, bracketed identity strings of the types
included in this argument will be ignored.}
}
\value{
An updated metabCombiner object containing a table of ordered feature
pairs for fitting a nonlinear rt fitting model. The columns are as follows:

\item{idx}{Identities of features from dataset X}
\item{idy}{Identities of features from dataset Y}
\item{mzx}{m/z values of features from dataset X}
\item{mzy}{m/z values of features from dataset Y}
\item{rtx}{retention time values of features from dataset X}
\item{rty}{retention time values of features from dataset Y}
\item{Qx}{abundance quantile values of features from dataset X}
\item{Qy}{abundance quantile values of features from dataset Y}
\item{adductX}{adduct label of features from dataset X}
\item{adductY}{adduct label of features from dataset Y}
\item{group}{m/z feature group of feature pairing}
\item{labels}{Anchor labels; "I" for identity, "A" for normal anchors}
}
\description{
A subset of all possible alignments in the \code{combinedTable} are used as
ordered pairs to anchor a retention time projection model. Alignments of
abundant features are prominent targets for anchor selection, but shared
identified metabolites may also be used.
}
\details{
In order to map between two sets of retention times, a set of ordered pairs
need to be selected for the spline fit. This function relies on mutually
abundant features to select these ordered pairs. In iterative steps, the most
abundant (as indicated by Q value) in one dataset is selected along with its
counterpart, and all features within some retention time window specified by
\code{windX} & \code{windY} arguments are excluded. This process is repeated
until all features have been considered. \code{tolQ} & \code{tolMZ} arguments
can restrict to pairs that have differences in Q & m/z within these tolerances.

Shared identities (in which idx & idy columns have matching, non-empty &
non-bracketed strings) may be used if \code{useID} ise set to \code{TRUE}. In
this case, shared identities will be searched first and will not be subject
to any of the restrictions in m/z, Q, or rt. The iterative process proceeds
after processing of shared identities.
}
\examples{
\dontrun{
library(metabCombiner)
data(plasma30)
data(plasma20)

p30 <- metabData(plasma30, samples = "CHEAR")
p20 <- metabData(plasma20, samples = "Red", rtmax = 17.25)
p.combined = metabCombiner(xdata = p30, ydata = p20, binGap = 0.0075)

##example 1 (no known IDs used)
p.combined = selectAnchors(p.combined, tolMZ = 0.003, tolQ = 0.3, windX = 0.03,
                           windY = 0.02)

##example 2 (known IDs used)
p.combined = selectAnchors(p.combined, useID = TRUE, tolMZ = 0.003, tolQ = 0.3)

##To View Plot of Ordered Pairs
anchors = getAnchors(p.combined)
plot(anchors$rtx, anchors$rty, main = "Selected Anchor Ordered Pairs",
     xlab = "rtx", ylab = "rty")
}
}
