% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2file.R
\name{write2file}
\alias{write2file}
\title{Print a Combiner Report to file.}
\usage{
write2file(object, file = "", sep = ",")
}
\arguments{
\item{object}{metabCombiner object or combinerTable.}

\item{file}{character string naming an output file path.}

\item{sep}{Character field separator. Values within each row are separated
by this string.}
}
\description{
Prints a \code{combinerTable} report to a file, specified by
             \code{file} argument. The output will contain an empty line
             separating each m/z group.
}
